+++ lib/application/workout/handlers/generate_handler.dart
@@ -1,6 +1,5 @@
 import 'package:dartz/dartz.dart';
 import 'package:flutter/foundation.dart';
-import 'package:bloc/bloc.dart';
 
 import 'package:power_progress/domain/exercise/exercise.dart';
 import 'package:power_progress/domain/exercise/i_exercise_repository.dart';