+++ lib/presentation/pages/onboarding/onboarding_informations_page.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 import 'package:power_progress/presentation/widgets/centered_loading.dart';
-import 'package:power_progress/presentation/theme/pp_theme.dart';
+import 'package:power_progress/presentation/theme/gradients.dart';
 import 'package:power_progress/application/exercise/exercise_cubit.dart';
 import 'package:power_progress/application/onboarding/onboarding_cubit.dart';
 import 'package:power_progress/core/util/spacing.dart';
@@ -32,7 +32,7 @@ class OnboardingInformationsPage extends StatelessWidget {
     return Scaffold(
       body: Container(
         decoration: BoxDecoration(
-          gradient: PPTheme.royalBlueGradient,
+          gradient: royalBlueGradient,
         ),
         child: BlocConsumer<ExerciseCubit, ExerciseState>(
           listener: (BuildContext context, ExerciseState state) {