+++ lib/presentation/theme/gradients.dart
@@ -1,0 +1,143 @@
+import 'package:flutter/material.dart';
+import 'package:power_progress/presentation/theme/pp_theme.dart';
+
+LinearGradient pinkYellowGradient = const LinearGradient(
+  colors: [
+    PPTheme.pink,
+    PPTheme.yellow,
+  ],
+  begin: Alignment.topLeft,
+  end: Alignment.bottomRight,
+  stops: [0.3, 1],
+);
+
+LinearGradient purplePinkGradient = const LinearGradient(
+  colors: [
+    Color(0xFF654ea3),
+    Color(0xFFeaafc8),
+  ],
+  begin: Alignment.topLeft,
+  end: Alignment.bottomRight,
+  stops: [0.3, 1],
+);
+
+LinearGradient redBlackGradient = const LinearGradient(
+  colors: [
+    Colors.black,
+    PPTheme.red,
+  ],
+  begin: Alignment.topLeft,
+  end: Alignment.bottomRight,
+  stops: [0.3, 1],
+);
+
+LinearGradient royalBlueGradient = const LinearGradient(
+  colors: [
+    Color(0xFF292E49),
+    Color(0xFF536976),
+  ],
+  begin: Alignment.topLeft,
+  end: Alignment.bottomRight,
+  stops: [0.3, 1],
+);
+
+LinearGradient burningOrangeGradient = const LinearGradient(
+  colors: [
+    Color(0xFFFF4B2B),
+    Color(0xFFFF416C),
+  ],
+  begin: Alignment.topLeft,
+  end: Alignment.bottomRight,
+  stops: [0, 1],
+);
+
+LinearGradient citrusPeelGradient = const LinearGradient(
+  colors: [
+    Color(0xFFFDC830),
+    Color(0xFFF37335),
+  ],
+  begin: Alignment.topLeft,
+  end: Alignment.bottomRight,
+  stops: [0, 1],
+);
+
+LinearGradient blueRaspberryGradient = const LinearGradient(
+  colors: [
+    Color(0xFF00B4DB),
+    Color(0xFF0083B0),
+  ],
+  begin: Alignment.topLeft,
+  end: Alignment.bottomRight,
+  stops: [0, 1],
+);
+
+LinearGradient ohhappinessGradient = const LinearGradient(
+  colors: [
+    Color(0xFF00b09b),
+    Color(0xFF96c93d),
+  ],
+  begin: Alignment.topLeft,
+  end: Alignment.bottomRight,
+  stops: [0, 1],
+);
+
+LinearGradient orcaGradient = const LinearGradient(
+  colors: [
+    Color(0xFF093637),
+    Color(0xFF093637),
+    Color(0xFF44A08D),
+  ],
+  begin: Alignment.bottomRight,
+  end: Alignment.topLeft,
+  stops: [0, 0.7, 1],
+);
+
+LinearGradient haikusGradient = const LinearGradient(
+  colors: [
+    Color(0xFFfd746c),
+    Color(0xFFff9068),
+  ],
+  begin: Alignment.topLeft,
+  end: Alignment.bottomRight,
+  stops: [0, 1],
+);
+
+LinearGradient cherryGradient = const LinearGradient(
+  colors: [
+    Color(0xFFEB3349),
+    Color(0xFFF45C43),
+  ],
+  begin: Alignment.topLeft,
+  end: Alignment.bottomRight,
+  stops: [0, 1],
+);
+
+LinearGradient juicyOrangeGradient = const LinearGradient(
+  colors: [
+    Color(0xFFFF8008),
+    Color(0xFFFFC837),
+  ],
+  begin: Alignment.topLeft,
+  end: Alignment.bottomRight,
+  stops: [0, 1],
+);
+
+LinearGradient namnGradient = const LinearGradient(
+  colors: [
+    Color(0xFFa73737),
+    Color(0xFF7a2828),
+  ],
+  begin: Alignment.topLeft,
+  end: Alignment.bottomRight,
+  stops: [0, 1],
+);
+
+LinearGradient lightGradient = const LinearGradient(
+  colors: [
+    Color(0x00000000),
+    Color(0xFFeeeeee),
+  ],
+  begin: Alignment.bottomRight,
+  end: Alignment.topLeft,
+  stops: [0.5, 1],
+);