+++ CHANGELOG.md
@@ -2,6 +2,36 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [2.42.1](https://github.com/KiritchoukC/power_progress/compare/v2.42.0...v2.42.1) (2020-08-06)
+
+## [2.42.0](https://github.com/KiritchoukC/power_progress/compare/v2.41.0...v2.42.0) (2020-08-06)
+
+
+### Features
+
+* **predefined-exercises:** update onboarding flow to include exercise picking and finish ui ([58d5951](https://github.com/KiritchoukC/power_progress/commit/58d5951)), closes [#76](https://github.com/KiritchoukC/power_progress/issues/76)
+
+## [2.41.0](https://github.com/KiritchoukC/power_progress/compare/v2.40.0...v2.41.0) (2020-08-05)
+
+
+### Features
+
+* **predefined-exercises:** continue on ui ([a11bf15](https://github.com/KiritchoukC/power_progress/commit/a11bf15)), closes [#76](https://github.com/KiritchoukC/power_progress/issues/76)
+
+## [2.40.0](https://github.com/KiritchoukC/power_progress/compare/v2.39.0...v2.40.0) (2020-08-03)
+
+
+### Features
+
+* **predefined-exercises:** continue on exercise picker ui ([cc32b87](https://github.com/KiritchoukC/power_progress/commit/cc32b87)), closes [#76](https://github.com/KiritchoukC/power_progress/issues/76)
+
+## [2.39.0](https://github.com/KiritchoukC/power_progress/compare/v2.38.14...v2.39.0) (2020-07-30)
+
+
+### Features
+
+* **predefined-exercises:** start exercise picker ([cbc0923](https://github.com/KiritchoukC/power_progress/commit/cbc0923)), closes [#76](https://github.com/KiritchoukC/power_progress/issues/76)
+
 ### [2.38.14](https://github.com/KiritchoukC/power_progress/compare/v2.38.13...v2.38.14) (2020-07-29)